ref : 쿠폰의 성능 개선을 위한 기능들을 test 패키지로 분류, 불필요 파일 삭제 및 jacoco에서 인식 못하도록 gradle 설정 추가
