+++ athena/build.gradle
@@ -133,6 +133,7 @@ jacocoTestReport {
 					'**/controller/**',
 					'**/mapper/**',
 					'**/repository/**',
+					'**/service/test/**',
 					'**/dto/**',
 					'**/*Application*',
 					'**/exception/**',
@@ -179,6 +180,7 @@ jacocoTestCoverageVerification {
 					'**/p6spy/**',
 					'**/controller/**',
 					'**/repository/**',
+					'**/service/test/**',
 					'**.dto.**',
 					'**.*Application*',
 					'**.exception.**',
